[BE]feat: 회원 대시보드 조회 기능 구현

+++ backend/src/main/java/com/tunapearl/saturi/dto/user/UserStreakInfoDaysDTO.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @AllArgsConstructor
 public class UserStreakInfoDaysDTO {
     private UserStreakDateDTO streakDate;
-    private Long solvedNum;
+    private Integer solvedNum;
 }
