+++ backend/src/main/java/com/tunapearl/saturi/dto/user/UserTotalLessonInfoDTO.java
@@ -3,9 +3,11 @@ package com.tunapearl.saturi.dto.user;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.util.Objects;
+
 @Getter
 @AllArgsConstructor
 public class UserTotalLessonInfoDTO {
-    private Long totalLessonGroup;
-    private Long totalLesson;
+    private int totalLessonGroup;
+    private int totalLesson;
 }
